pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.66.3",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "streamlit>=1.43.2",
]
